+++ src/io.rs
@@ -3,7 +3,33 @@
 #[cfg(feature = "biblatex")]
 use biblatex::{Bibliography, TypeError};
 
-use crate::Entry;
+use crate::{Entry, Library};
+
+/// Parse a bibliography from a YAML string.
+///
+/// ```
+/// use hayagriva::io::from_yaml_str;
+///
+/// let yaml = r#"
+/// crazy-rich:
+///     type: Book
+///     title: Crazy Rich Asians
+///     author: Kwan, Kevin
+///     date: 2014
+///     publisher: Anchor Books
+///     location: New York, NY, US
+/// "#;
+/// let bib = from_yaml_str(yaml).unwrap();
+/// assert_eq!(bib.nth(0).unwrap().date().unwrap().year, 2014);
+/// ```
+pub fn from_yaml_str(s: &str) -> Result<Library, serde_yaml::Error> {
+    serde_yaml::from_str(s)
+}
+
+/// Serialize a bibliography to a YAML string.
+pub fn to_yaml_str(entries: &Library) -> Result<String, serde_yaml::Error> {
+    serde_yaml::to_string(&entries)
+}
 
 /// Errors that may occur when parsing a BibLaTeX file.
 #[cfg(feature = "biblatex")]
@@ -54,5 +80,27 @@ pub fn from_biblatex(bibliography: &Bibliography) -> Result<Vec<Entry>, Vec<Type
         Err(errors)
     } else {
         Ok(res.into_iter().map(|item| item.unwrap()).collect())
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use std::fs;
+
+    #[test]
+    fn roundtrip() {
+        let contents = fs::read_to_string("tests/basic.yml").unwrap();
+        let entries = from_yaml_str(&contents).unwrap();
+        let yaml = to_yaml_str(&entries).unwrap();
+        println!("{}", &yaml);
+
+        let reconstructed = from_yaml_str(&yaml).unwrap();
+        assert_eq!(entries.len(), reconstructed.len());
+
+        for entry in entries {
+            let match_e = reconstructed.iter().find(|x| x.key == entry.key).unwrap();
+            assert_eq!(match_e, &entry);
+        }
     }
 }
